[Editor] Organize folders and namespaces
