+++ MC/McEditor/App.xaml
@@ -4,7 +4,7 @@
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:local="clr-namespace:McEditor"
-             StartupUri="Main/MainWindow.xaml">
+             StartupUri="Windows/Main/MainWindow.xaml">
     <Application.Resources>
 
         <ResourceDictionary>
