+++ MC/McEditor/Controls/Canvas/CanvasControl.xaml
