+++ MC/McEditor/Controls/Canvas/CanvasControlViewModel.cs
