+++ MC/McEditor/Controls/Canvas/ICanvasControl.cs
