+++ MC/McEditor/Service/ResourceService.cs
@@ -9,7 +9,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace McEditor
+namespace McEditor.Service
 {
     public class ResourceService : NotifyPropertyChangedBase
     {
