+++ MC/McEditor/Windows/Main/IMainWindow.cs
@@ -7,7 +7,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace McEditor
+namespace McEditor.Windows
 {
     internal interface IMainWindow
     {
