+++ MC/McEditor/Windows/Main/MainWindow.xaml
@@ -1,0 +1,38 @@
+﻿<!--Copyright (c) 2022 K.T
+ Licensed under the MIT License. See LICENSE in the project root for license information.-->
+
+<Window x:Class="McEditor.Windows.MainWindow"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:local="clr-namespace:McEditor.Windows"
+        xmlns:service="clr-namespace:McEditor.Service"
+        mc:Ignorable="d"
+        Title="MainWindow" Height="450" Width="800" Background="{DynamicResource CustomWindowBackground}" WindowStartupLocation="CenterScreen" WindowState="Maximized" MinWidth="800" MinHeight="450">
+
+    <Window.DataContext>
+        <local:MainWindowViewModel/>
+    </Window.DataContext>
+
+    <Grid>
+        <Menu Height="22" VerticalAlignment="Top" d:LayoutOverrides="Height" Background="White">
+            <MenuItem Header="{Binding Source={x:Static service:ResourceService.Current}, Path=Resources.File, Mode=OneWay}" Height="22" Width="60">
+                <MenuItem Header="New Canvas"/>
+            </MenuItem>
+            <MenuItem  Header="{Binding Source={x:Static service:ResourceService.Current}, Path=Resources.Edit, Mode=OneWay}" Height="22" Width="50"/>
+            <MenuItem Header="{Binding Source={x:Static service:ResourceService.Current}, Path=Resources.Debug, Mode=OneWay}"  Height="22" Width="70"/>
+            <MenuItem Header="{Binding Source={x:Static service:ResourceService.Current}, Path=Resources.Window, Mode=OneWay}"  Height="22" Width="80"/>
+            <MenuItem Header="{Binding Source={x:Static service:ResourceService.Current}, Path=Resources.Developer, Mode=OneWay}"  Height="22" Width="70">
+                <MenuItem Header="{Binding Source={x:Static service:ResourceService.Current}, Path=Resources.Language, Mode=OneWay}">
+                    <MenuItem Header="日本語" Command="{Binding ChangeLanguageCommand,Mode=OneWay,UpdateSourceTrigger=PropertyChanged}"
+                               CommandParameter="{x:Static service:ResourceService.JAPANACECODE}"/>
+                    <MenuItem Header="English" Command="{Binding ChangeLanguageCommand,Mode=OneWay,UpdateSourceTrigger=PropertyChanged}"
+                               CommandParameter="{x:Static service:ResourceService.ENGLISHCODE}"/>
+                </MenuItem>
+            </MenuItem>
+            <MenuItem Header="{Binding Source={x:Static service:ResourceService.Current}, Path=Resources.Help, Mode=OneWay}" Height="22" Width="60"/>
+        </Menu>
+
+    </Grid>
+</Window>
