+++ MC/McEditor/Windows/Main/MainWindow.xaml.cs
@@ -17,7 +17,7 @@ using System.Windows.Media.Imaging;
 using System.Windows.Navigation;
 using System.Windows.Shapes;
 
-namespace McEditor
+namespace McEditor.Windows
 {
     /// <summary>
     /// Interaction logic for MainWindow.xaml
